requirements.txt
alembic==1.13.1
amqp==5.2.0
aniso8601==9.0.1
APScheduler==3.10.4
asttokens==2.4.1
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
billiard==4.2.0
black==24.2.0
blinker==1.7.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.1
cryptography==38.0.4
debugpy==1.8.1
decorator==5.1.1
dicebear==2.7.2
elastic-transport==8.12.0
elasticsearch==8.12.1
exceptiongroup==1.2.0
executing==2.0.1
factory-boy==3.3.0
Faker==24.1.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
flask-pyjwt==1.0.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
importlib_resources==6.1.3
iniconfig==2.0.0
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kombu==5.3.6
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==23.2
pandas==2.2.1
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.1
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.28
stack-data==0.6.3
tomli==2.0.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
